+++ Example/Tests/DezignableRotation.swift
@@ -5,10 +5,8 @@ import Dezignables
 class DezignableRotationTests: Base {
   
   func testRotation() {
-    let dezignableView = DezignableView(frame: CGRectMake(0, 0, 200, 200))
-    dezignableView.backgroundColor = UIColor.redColor()
-    dezignableView.rotation = 45
-    FBSnapshotVerifyView(dezignableView)
+    self.subject.rotation = 45
+    FBSnapshotVerifyView(self.subject)
   }
   
 }