+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added `cornerCircle` to create a full circle based on the height of the view
 - Added more examples and documentation
 - Added `DezignableBackground` with background fill color and gradient background
+- Added `DezignableShadow`
 
 ## 1.1.0 - 26 feb 2016
 - Mostly refactored to protocol based extensions