+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
 
-## 1.2.0 - master
+## 1.3.0 - master
+
+## 1.2.0 - 1 Mar 2016
 - Added `cornerCircle` to create a full circle based on the height of the view
 - Added more examples and documentation
 - Added `DezignableBackground` with background fill color and gradient background
 - Added `DezignableShadow`
 - Added component `DezignableLabel`
+- Added component `DezignableTableViewController`
 
-## 1.1.0 - 26 feb 2016
+## 1.1.0 - 26 Feb 2016
 - Mostly refactored to protocol based extensions
 - Added `DezignableRotation`
 
-## 1.0.0 - 26 feb 2016
+## 1.0.0 - 26 Feb 2016
 - Initial Release